# swagger/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GauthierFitness \u2014 API Documentation",
        "version": "1.0.0",
        "description": "**API REST de GauthierFitness** \u2014 boutique e-commerce de produits fitness personnalisables (v\u00eatements, accessoires, nutrition).\n\n## Conventions\n- Authentification : **Laravel Sanctum** (`Bearer <token>`). Le token est obtenu via `POST /api/login` ou `POST /api/register`.\n- Format : JSON (`Accept: application/json`).\n- Codes d'erreur :\n  - `401` \u2014 non authentifi\u00e9\n  - `403` \u2014 interdit (ex : ressource d'un autre utilisateur)\n  - `404` \u2014 ressource introuvable\n  - `422` \u2014 donn\u00e9es de validation invalides\n- Rate limiting : `POST /api/contact` est limit\u00e9 \u00e0 5 requ\u00eates/minute/IP.\n\n## Domaines fonctionnels\n- **Auth** \u2014 inscription, connexion, d\u00e9connexion, v\u00e9rification.\n- **Catalogue** \u2014 produits publics (listing pagin\u00e9 + d\u00e9tail).\n- **Panier & Commande** \u2014 gestion du panier, paiement Stripe, suivi commande.\n- **Customisation** \u2014 sessions de personnalisation 2D/3D + assets (logos, images).\n- **IA** \u2014 g\u00e9n\u00e9ration d'images via OpenAI pour designs personnalis\u00e9s.\n- **Admin** \u2014 produits, commandes, stocks (r\u00e9serv\u00e9 aux administrateurs)."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api"
        }
    ],
...
    "security": [
        {
            "Sanctum": []
        }
    ],
    "paths": {
...
        "/customization/assets/logo": {
            "post": {
                "operationId": "customizationAsset.uploadLogo",
                "summary": "Upload a logo (PNG/JPG/JPEG/WebP, \u2264 3 MB).\nStores the file in `storage/app/public/customization/logos/{user_id}/`",
                "tags": [
                    "Customisation"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 422
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/customization/assets/image": {
            "post": {
                "operationId": "customizationAsset.uploadImage",
                "summary": "Upload a customization image (PNG/JPG/JPEG/WebP, \u2264 5 MB).\nStores the file in `storage/app/public/customization/images/{user_id}/`",
                "tags": [
                    "Customisation"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "integer",
                                    "const": 422
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/products/{product}/stock/movements": {
            "get": {
                "operationId": "adminStock.movements",
                "summary": "Paginated history of a product's stock movements.\nReturns all `stock_movements` (in, out, corrections) linked to the product, 30 per page, most recent first",
                "tags": [
                    "Admin \u2014 Stock"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "minimum": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/StockMovement"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "from": {
                                            "type": [
                                                "integer",
                                                "null"
                                            ],
                                            "minimum": 1
                                        },
                                        "last_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "minimum": 1
                                        },
                                        "links": {
                                            "type": "array",
                                            "description": "Generated paginator links.",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "url": {
                                                        "type": [
                                                            "string",
                                                            "null"
                                                        ]
                                                    },
                                                    "label": {
                                                        "type": "string"
                                                    },
                                                    "active": {
                                                        "type": "boolean"
                                                    }
                                                },
                                                "required": [
                                                    "url",
                                                    "label",
                                                    "active"
                                                ]
                                            }
                                        },
                                        "next_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "path": {
                                            "type": [
                                                "string",
                                                "null"
                                            ],
                                            "description": "Base path for paginator generated URLs."
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "description": "Number of items shown per page.",
                                            "minimum": 0
                                        },
                                        "prev_page_url": {
                                            "type": [
                                                "string",
                                                "null"
                                            ]
                                        },
                                        "to": {
                                            "type": [
                                                "integer",
                                                "null"
                                            ],
                                            "description": "Number of the last item in the slice.",
                                            "minimum": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "description": "Total number of items being paginated.",
                                            "minimum": 0
                                        }
                                    },
                                    "required": [
                                        "current_page",
                                        "data",
                                        "first_page_url",
                                        "from",
                                        "last_page_url",
                                        "last_page",
                                        "links",
                                        "next_page_url",
                                        "path",
                                        "per_page",
                                        "prev_page_url",
                                        "to",
                                        "total"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Sanctum": {
                "type": "http",
                "description": "Token Sanctum obtenu via `POST /api/login` ou `POST /api/register`. Pr\u00e9fix\u00e9 par `Bearer ` dans le header `Authorization`.",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "StockMovement": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "lot_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    },
                    "type": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "reason": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "format": "date-time"
                    },
                    "product_option_id": {
                        "type": [
                            "integer",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "lot_id",
                    "product_id",
                    "user_id",
                    "type",
                    "quantity",
                    "reason",
                    "created_at",
                    "updated_at",
                    "product_option_id"
                ],
                "title": "StockMovement"
            },
...
        },
        "responses": {
...
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}